requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
